Move GRUB2 to pyanaconda.bootloader.grub2
Move the GRUB2 class to a new file.
